# contracts/openapi.json
{"openapi":"3.1.0","info":{"title":"TouchMap preparation API","version":"1.0.0"},"paths":{"/health/live":{"get":{"summary":"Live","operationId":"live_health_live_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health/ready":{"get":{"summary":"Ready","operationId":"ready_health_ready_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/sessions/pair":{"post":{"summary":"Pair","operationId":"pair_v1_sessions_pair_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PairRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/sessions/current":{"delete":{"summary":"Revoke","operationId":"revoke_v1_sessions_current_delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/jobs/{request_id}":{"get":{"summary":"Get Job","operationId":"get_job_v1_jobs__request_id__get","parameters":[{"name":"request_id","in":"path","required":true,"schema":{"type":"string","title":"Request Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Cancel Job","operationId":"cancel_job_v1_jobs__request_id__delete","parameters":[{"name":"request_id","in":"path","required":true,"schema":{"type":"string","title":"Request Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/import/svg":{"post":{"summary":"Import Svg","operationId":"import_svg_v1_import_svg_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_import_svg_v1_import_svg_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/analyze/image":{"post":{"summary":"Analyze","operationId":"analyze_v1_analyze_image_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_analyze_v1_analyze_image_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/audio/labels":{"post":{"summary":"Speech","operationId":"speech_v1_audio_labels_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpeechRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/packages/validate":{"post":{"summary":"Validate","operationId":"validate_v1_packages_validate_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_validate_v1_packages_validate_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_analyze_v1_analyze_image_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"},"draftRevision":{"type":"integer","title":"Draftrevision"},"language":{"type":"string","title":"Language"},"requestId":{"type":"string","title":"Requestid"},"sourceHash":{"type":"string","title":"Sourcehash"},"consent":{"type":"boolean","title":"Consent"}},"type":"object","required":["file","draftRevision","language","requestId","sourceHash","consent"],"title":"Body_analyze_v1_analyze_image_post"},"Body_import_svg_v1_import_svg_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"},"draftRevision":{"type":"integer","title":"Draftrevision"},"language":{"type":"string","title":"Language"},"requestId":{"type":"string","title":"Requestid"}},"type":"object","required":["file","draftRevision","language","requestId"],"title":"Body_import_svg_v1_import_svg_post"},"Body_validate_v1_packages_validate_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"},"requestId":{"type":"string","title":"Requestid"}},"type":"object","required":["file","requestId"],"title":"Body_validate_v1_packages_validate_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LabelRequest":{"properties":{"targetId":{"type":"string","maxLength":128,"minLength":1,"pattern":"^[A-Za-z0-9][A-Za-z0-9_.-]*$","title":"Targetid"},"kind":{"type":"string","pattern":"^(label|description|question|overview)$","title":"Kind"},"text":{"type":"string","maxLength":4000,"minLength":1,"title":"Text"},"textHash":{"type":"string","pattern":"^[a-f0-9]{64}$","title":"Texthash"}},"additionalProperties":false,"type":"object","required":["targetId","kind","text","textHash"],"title":"LabelRequest"},"PairRequest":{"properties":{"code":{"type":"string","maxLength":128,"minLength":6,"title":"Code"}},"additionalProperties":false,"type":"object","required":["code"],"title":"PairRequest"},"SpeechRequest":{"properties":{"requestId":{"type":"string","maxLength":128,"minLength":1,"pattern":"^[A-Za-z0-9][A-Za-z0-9_.-]*$","title":"Requestid"},"sourceHash":{"type":"string","pattern":"^[a-f0-9]{64}$","title":"Sourcehash"},"draftRevision":{"type":"integer","minimum":1.0,"title":"Draftrevision"},"language":{"type":"string","maxLength":35,"pattern":"^[a-zA-Z]{2,3}(?:-[a-zA-Z0-9]{2,8})*$","title":"Language"},"consent":{"type":"boolean","title":"Consent"},"labels":{"items":{"$ref":"#/components/schemas/LabelRequest"},"type":"array","maxItems":100,"minItems":1,"title":"Labels"}},"additionalProperties":false,"type":"object","required":["requestId","sourceHash","draftRevision","language","consent","labels"],"title":"SpeechRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}